Manche Komponenten dürfen nicht mehr entfernt werden

+++ src/scenes/SettingScene.ts
@@ -111,16 +111,19 @@ export default class SettingScene extends SidebarScene {
         .dom(0, 0, 'h3', '', `${component.name} (ID: ${component.id})`)
         .setClassName('componentTitle');
 
-      const deleteButton = this.add.dom(0, 0, 'div', '', '✖').setClassName('deleteButton');
-      deleteButton.setData('ignoreHeight', true);
-      deleteButton.addListener('click');
-      deleteButton.on('click', () => {
-        EntityManager.removeComponent(entity.id, component);
-        // alle Componenten der Enittät neu laden
-        container.removeAll(true);
-        container.height = 0;
-        this.onCreate(container, entity);
-      });
+      let deleteButton;
+      if (component.isDeletable()) {
+        deleteButton = this.add.dom(0, 0, 'div', '', '✖').setClassName('deleteButton');
+        deleteButton.setData('ignoreHeight', true);
+        deleteButton.addListener('click');
+        deleteButton.on('click', () => {
+          EntityManager.removeComponent(entity.id, component);
+          // alle Componenten der Enittät neu laden
+          container.removeAll(true);
+          container.height = 0;
+          this.onCreate(container, entity);
+        });
+      }
 
       const attributes = Object.keys(component).map(attribute => {
         if (component[attribute] instanceof Attribute) {
